fix: Move darkmodeJs into _document.tsx
